Refactor ServerCommandHandler
Now initaliation of each command is
hidden inside ServerCommandHandler.
