--- server/src/channels/components/mod.rs
@@ -1,2 +1,0 @@
-pub mod messages_cleaner;
-pub mod messages_saver;
