+++ server/src/channels/mod.rs
@@ -1,4 +1,3 @@
 pub mod commands;
-pub mod components;
-pub mod executor;
 pub mod handler;
+pub mod server_command;
